CP-28298: Finish implementing time based authenticated write
Implement time based authenticated writes to be fully compliant (modulo
bugs) with the UEFI 2.7 A specification. Much of the code is adapted
from the edk2.

This code is enough for a working implementation of Secure Boot and to
pass the Firmware Test Suite's secure boot tests.

Also create some expected variables related to SecureBoot.