CP-28669: Load keys on firstboot
When a VM is booted for the first time, populate PK, KEK, and db.

Appropriate key(s) for KEK and db need to be generated or acquired and
placed into KEK.list and db.list respectively. Keys can be generated
with the default Makefile rule by running, e.g. make db.pem