CP-28668: Add command-line tools for interacting with EFI variables
Add some command-line tools for interacting with a guest's EFI variables
when the guest is not running. This may allow a VM to be recovered in
certain situations and is also useful when debugging.

* varstore-ls - Lists a VM's variables
* varstore-get - Show attributes and contents of a variable
* varstore-rm - Removes a variable
* varstore-set - Sets, updates, appends, or removes a new variable
  depending on the specified attributes. Authenticated variables are
  supported.
* varstore-sb-state - Sets or clears the variables related to Secure
  Boot.

As part of this, add a new xapidb-cmdline backend which shares most of
the code with the existing xapidb backend. Move common headers into the
include directory.