CP-29064: Fetch blob directly from XAPI at startup
If xenopsd writes out the init blob, there is a race condition because
during a guest reboot xenopsd might not have the most up-to-date
information from XAPI and so it may use stale data. So just fetch the
data directly from XAPI. This also allows more code to be shared with
the xapidb-cmdline backend.