CP-29171/CA-295626: Add extra checks to verify_auth_var_type
Add some checks that are needed and remove some checks that are not
needed due to certain invariants. Check that the digest used is always
SHA-256 and enable the corresponding previously failing test. Refactor
the code so that signatures are wrapped earlier if needed which makes
the code for checking the digest simpler and makes other parts of the
code simpler and faster. Only load the single supported digest and move
the loading of it to the start-of-day. Also change the error strings to
be loaded only when needed and freed after use.