CP-29373: Throttle writes to XAPI
Throttle writes to XAPI to avoid the guest DoSing XAPI. It uses a
variation of the leaky bucket algorithm where instead of leaking, it
waits for credit before completing the request. The result is that
guests that write variables intermittently will never be throttled.
Guests that write many times in a short space of time will be throttled,
but not permanently. This allows a sustained write speed of twice per
second.