CP-29551: Restrict Xen handles
Restrict the Xen handles so that varstored can only act on a single
domain. This requires two large changes:

* Switching to use the newer Xen libraries and APIs rather than using
the older, compat Xen APIs which don't allow being restricted.
* Removing the sequence-based initialization and teardown. The
sequence-based initialization and teardown had a lot of boilerplate code
and complexity without serving much useful purpose. Replace it with a
much shorter and simpler initialization and teardown which allows the
order of some code to change around as required when restricting Xen
handles.