Set an infinite poll timeout
Set an infinite poll timeout to prevent spurious wakeups. Poll will
wakeup if either a file descriptor is ready, a signal has occurred, or a
timeout has occurred. We are interested in the first two kinds of wakeup
but not the latter. There is no reason for varstored to wakeup on
timeout since it just reenters poll. Therefore, make the timeout
infinite.