Cleanup initialize_settings()
Remove the feature flag toggle since the feature flag design states that
flags should be removed once before release. It also states that feature
flags should default to off (except on a feature branch) whereas the
code defaulted to on if the feature flag was not present.

Remove the weird all-caps logging and change it to be more accurate.
i.e. that secure_boot_enable is true, not that secure boot is in
operation since that depends on the contents of a guest's variables.

Remove a useless comment.