CA-333944: Set SignatureOwner to Microsoft's GUID for KEK and db
A signature database (EFI_SIGNATURE_DATA) has a SignatureOwner field
defined as follows:

  An identifier which identifies the agent which added the signature to
  the list.

However, the Microsoft Windows Hardware Compatibility Specifications
require that the Signature owner for its signature database in KEK and
db is a particular Microsoft GUID (and this is tested by
Microsoft.UefiSecureBootLogo.Tests.OutOfBoxVerifyMicrosoftKEKpresent).
This is contrary to the UEFI spec but since it probably doesn't matter
either way, use the Microsoft GUID so that the test is happy.