CP-35896: Add a coverity model
Add a model for asprintf and vasprintf from blktap so that Coverity
handles them properly.
Add a model for xmlXPathFreeObject so that Coverity knows that the
object has been freed.
Define dummy logging macros for Coverity so that it doesn't complain
about dead code since the log level is defined at compile time.