CP-35896: Fix loading the certificate digest from serialized state
Fix a missing dereference which meant that the certificate digest was
incorrectly loaded from the serialized state. This meant that it was not
possible to change or remove an authenticate variable with
type == AUTH_TYPE_PRIVATE from within the guest after a reboot or
suspend/resume cycle.

Found by Coverity - CID 90956

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>