CA-351587: Move crypto initialization earlier
Crypto initialization was previously done in setup_variables() but this
path is only called during the VM boot, not during resume. This meant
that trying to set an authenticated variable after resume would fail
with a PKCS7_R_UNKNOWN_DIGEST_TYPE error.

Move it earlier so it is called in both cases.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>