Switch to using new xendevicemodel_nr_vcpus() API
This is a stable API replacing the need to use the unstable
xc_domain_getinfo().  This in turn drops the dependency on libxenctrl.

As part of dropping the xenctrl.h include, include xen/memory.h for the
resource mapping constants, and swap xc_evtchn_port_or_error_t for
xenevtchn_port_or_error_t which is provided by xenevtchn.h.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>