CA-352332: Skip setting PK if KEK/db are missing
The KEK and db auth contents may be missing during VM firstboot if they
could not be captured from the host.  In that case, PK would still be
set which transitions the firmware into user mode.

While in user mode, dbx updates need to be signed with a key in PK or
KEK. This causes the dbx update in Windows Update KB4535680 to fail
since the signature that it is signed with is not present in KEK.

This is undesirable (though completely spec compliant) behaviour. To
avoid it happening, if KEK/db are missing avoid setting PK which will
mean that the firmware stays in setup mode and updates to dbx do not
need to be signed by a key in PK/KEK.

Leaving the firmware in setup mode does open a potential in-VM DoS
vector - malware within the VM could set PK/KEK/db itself and prevent
the regular OS from booting. However, this is easy enough to recover
from and malware with that level of privilege could likely achieve the
same behaviour in other ways.

In case the host does have KEK/db auth contents available, the behaviour
is unchanged.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>