CA-366706: Don't block most signals
It is not necessary and a bit odd to block most signals. In particular,
the result may be undefined when certain signals like SIGSEGV are
generated while blocked. Instead of blocking signals, just ignore
SIGPIPE.

Secondly, if varstored gets SIGABRT, it means something has gone wrong
(e.g. heap corruption). Don't try to catch it, just let it exit and
generate a coredump.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>