CP-309775: Add new Microsoft certificates
The existing Microsoft KEK and db certificates expire in 2026 so add the
new 2023 certificates. Newly provisioned VMs will get both old and new
certificates so that they will boot regardless of which key they were
signed with.

The certificates are from
https://github.com/microsoft/secureboot_objects and converted to PEM
format using openssl.

This time round there is a certificate specifically for option ROMs but
I have not included it here since it should not be necessary for
XenServer VMs.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>