Use memory barriers defined by xen headers
This gives us the necessary definitions for arm64.

Those live in xen-barrier.h starting with 4.20, but we won't be able
to use it for some time.  Before this they live in xenctrl.h, but it
pulls too much stuff.  So this is a sync with the 4.20 headers
contents.

Since the code only uses mb, and no [rw]mb, the latter are not
imported.  The original macro name from the Xen header is preserved
for easier future synchro, so this adjusts the callsites accordingly.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>